refactor: Game 클래스 로직을 Board로 이전 후 프로퍼티에 RenjuRuleAdapter 추가, lastStone을 Stones 클래스로 이동
