+++ src/main/kotlin/omok/model/Board.kt
@@ -1,58 +1,45 @@
 package omok.model
 
-import omok.model.adapter.BlackRenjuRuleAdapter
-import omok.model.adapter.WhiteRenjuRuleAdapter
+import omok.model.adapter.RenjuRuleAdapter
 import omok.model.game.FoulCondition
+import omok.model.game.GameState
 import omok.model.stone.Stone
 import omok.model.stone.StoneColor
 import omok.model.stone.Stones
 
-sealed class FoulConditionResult
-
 class Board(
-    private val blackStones: Stones = Stones(ruleAdapter = BlackRenjuRuleAdapter(MAX_BOARD_WIDTH, MAX_BOARD_HEIGHT)),
-    private val whiteStones: Stones = Stones(ruleAdapter = WhiteRenjuRuleAdapter(MAX_BOARD_WIDTH, MAX_BOARD_HEIGHT)),
+    val stones: Stones = Stones(),
+    private val renjuRuleAdapter: RenjuRuleAdapter,
 ) {
-    val stones: Set<Stone> get() = blackStones.stones + whiteStones.stones
-
     fun place(newStone: Stone) {
-        require(newStone.point.row in MIN_BOARD_HEIGHT..MAX_BOARD_HEIGHT && newStone.point.col in MIN_BOARD_WIDTH..MAX_BOARD_WIDTH) {
-            ERROR_MESSAGE_OUT_OF_BOARD
-        }
-        require(!stones.isOccupied(newStone)) { ERROR_MESSAGE_IS_ALREADY_OCCUPIED }
-        when (blackStones.checkAnyFoulCondition(whiteStones, newStone)) {
-            FoulCondition.DOUBLE_THREE -> throw IllegalArgumentException(ERROR_MESSAGE_DOUBLE_THREE_VIOLATION)
-            FoulCondition.DOUBLE_FOUR -> throw IllegalArgumentException(ERROR_MESSAGE_DOUBLE_FOUR_VIOLATION)
-            FoulCondition.OVERLINE -> throw IllegalArgumentException(ERROR_MESSAGE_OVERLINE_VIOLATION)
-            FoulCondition.NONE -> placeByColor(newStone)
-        }
-    }
-
-    fun hasOmok(newStone: Stone): Boolean =
-        when (newStone.color) {
-            StoneColor.WHITE -> whiteStones.checkWin(whiteStones, newStone)
-            StoneColor.BLACK -> blackStones.checkWin(blackStones, newStone)
-        }
-
-    private fun placeByColor(newStone: Stone) {
-        when (newStone.color) {
-            StoneColor.WHITE -> whiteStones.add(newStone)
-            StoneColor.BLACK -> blackStones.add(newStone)
+        when (foulCondition(stones, newStone)) {
+            FoulCondition.DOUBLE_THREE -> TODO()
+            FoulCondition.DOUBLE_FOUR -> TODO()
+            FoulCondition.OVERLINE -> TODO()
+            FoulCondition.NONE -> {
+                stones.add(newStone)
+                stones.setLastStone(newStone)
+            }
         }
     }
 
-    private fun Set<Stone>.isOccupied(newStone: Stone) = newStone.point in map { it.point }
+    fun gameState(newStone: Stone): GameState {
+        if (!renjuRuleAdapter.checkWin(stones.stones, newStone)) {
+            return GameState.PLAYING
+        }
+        return when (newStone.color) {
+            StoneColor.BLACK -> GameState.BLACK_OMOK
+            StoneColor.WHITE -> GameState.WHITE_OMOK
+        }
+    }
+
+    fun foulCondition(
+        stones: Stones,
+        newStone: Stone,
+    ): FoulCondition = renjuRuleAdapter.checkAnyFoulCondition(stones.stones, newStone)
 
     companion object {
-        const val MIN_BOARD_WIDTH = 1
-        const val MIN_BOARD_HEIGHT = 1
-        const val MAX_BOARD_WIDTH = 15
         const val MAX_BOARD_HEIGHT = 15
-
-        private const val ERROR_MESSAGE_OUT_OF_BOARD = "오목판 범위를 벗어났습니다."
-        private const val ERROR_MESSAGE_IS_ALREADY_OCCUPIED = "이미 돌이 있는 자리입니다."
-        private const val ERROR_MESSAGE_DOUBLE_THREE_VIOLATION = "삼삼 금수입니다."
-        private const val ERROR_MESSAGE_DOUBLE_FOUR_VIOLATION = "사사 금수입니다."
-        private const val ERROR_MESSAGE_OVERLINE_VIOLATION = "장목 금수입니다."
+        const val MAX_BOARD_WIDTH = 15
     }
 }
